+++ 14th-team5-iOS/Bibbi/Sources/Presentation/Camera/Reactor/ImageGenerateViewReactor.swift
@@ -19,6 +19,7 @@ public final class ImageGenerateViewReactor: Reactor {
     
     @Injected private var createAiImageGenerateUseCase: CreateAiImageGenerateUseCaseProtocol
     @Injected private var uploadAiImagePostUseCase: CreatePostUseCaseProtocol
+    @Injected private var provider: ServiceProviderProtocol
     @Navigator var imageGenerateNavigator: ImageGenerateNavigatorProtocol
     
     
@@ -88,6 +89,7 @@ public final class ImageGenerateViewReactor: Reactor {
                     if entity == nil {
                         return .just(.setError("업로드 중 오류가 발생했습니다. 다시 한 번 시도해주세요."))
                     }
+                    owner.provider.aiImageGlobalState.imageUploadDidFinish(success: true)
                     owner.imageGenerateNavigator.toHome()
                     return .empty()
                 }
