+++ 14th-team5-iOS/Core/Sources/Bibbi/BBServices/ServiceProvider.swift
@@ -22,6 +22,7 @@ public protocol ServiceProviderProtocol: AnyObject {
     var realEmojiGlobalState: RealEmojiGlobalStateType { get }
     var profilePageGlobalState: ProfileFeedGlobalStateType { get }
     
+    var aiImageGlobalState: AiImageGlobalStateType { get }
     var studioGlobalState: StudioGlobalStateType { get }
     var commentService: CommentServiceType { get }
 }
@@ -37,6 +38,7 @@ final public class ServiceProvider: ServiceProviderProtocol {
     public lazy var managementService: any ManagementServiceType = ManagementService(provider: self)
     
     public lazy var studioGlobalState: any StudioGlobalStateType = StudioGlobalState(provider: self)
+    public lazy var aiImageGlobalState: any AiImageGlobalStateType = AiImageGlobalState(provider: self)
     
     public lazy var commentService: any CommentServiceType = CommentService(provider: self)
